+++ modules/examples/swing/src/main/scala/moe/brianhsu/live2d/demo/swing/widget/SwingEffectSelector.scala
@@ -3,8 +3,8 @@ package moe.brianhsu.live2d.demo.swing.widget
 import moe.brianhsu.live2d.demo.app.DemoApp.{ClickAndDrag, FollowMouse}
 import moe.brianhsu.live2d.demo.swing.Live2DUI
 import moe.brianhsu.live2d.enitiy.avatar.effect.impl.{Breath, EyeBlink, FaceDirection}
-import moe.brianhsu.live2d.usecase.updater.impl.BasicUpdateStrategy
-import moe.brianhsu.live2d.usecase.updater.impl.BasicUpdateStrategy.EffectTiming.{AfterExpression, BeforeExpression}
+import moe.brianhsu.live2d.usecase.updater.impl.GenericUpdateStrategy
+import moe.brianhsu.live2d.usecase.updater.impl.GenericUpdateStrategy.EffectTiming.{AfterExpression, BeforeExpression}
 
 import java.awt.event.ActionEvent
 import java.awt.{GridBagConstraints, GridBagLayout}
@@ -92,8 +92,8 @@ class SwingEffectSelector(live2DWidget: Live2DUI) extends JPanel {
     }
   }
 
-  def syncWithStrategy(basicUpdateStrategy: BasicUpdateStrategy): Unit = {
-    val effects = basicUpdateStrategy.effects(BeforeExpression) ++ basicUpdateStrategy.effects(AfterExpression)
+  def syncWithStrategy(strategy: GenericUpdateStrategy): Unit = {
+    val effects = strategy.effects(BeforeExpression) ++ strategy.effects(AfterExpression)
     val hasEyeBlink = effects.exists(_.isInstanceOf[EyeBlink])
     val hasBreath = effects.exists(_.isInstanceOf[Breath])
     val hasFaceDirection = effects.exists(_.isInstanceOf[FaceDirection])