+++ modules/examples/swing/src/main/scala/moe/brianhsu/live2d/demo/swing/widget/SwingMotionSelector.scala
@@ -3,8 +3,8 @@ package moe.brianhsu.live2d.demo.swing.widget
 import moe.brianhsu.live2d.demo.swing.Live2DUI
 import moe.brianhsu.live2d.enitiy.avatar.Avatar
 import moe.brianhsu.live2d.enitiy.avatar.effect.impl.LipSyncFromMotionSound
-import moe.brianhsu.live2d.usecase.updater.impl.BasicUpdateStrategy
-import moe.brianhsu.live2d.usecase.updater.impl.BasicUpdateStrategy.EffectTiming.{AfterExpression, BeforeExpression}
+import moe.brianhsu.live2d.usecase.updater.impl.GenericUpdateStrategy
+import moe.brianhsu.live2d.usecase.updater.impl.GenericUpdateStrategy.EffectTiming.{AfterExpression, BeforeExpression}
 
 import java.awt.event.{ActionEvent, MouseAdapter, MouseEvent}
 import java.awt.{Component, GridBagConstraints, GridBagLayout}
@@ -94,8 +94,8 @@ class SwingMotionSelector(live2DWidget: Live2DUI) extends JPanel {
     expandAllNodes(motionTree, 0, motionTree.getRowCount)
   }
 
-  def syncWithStrategy(basicUpdateStrategy: BasicUpdateStrategy): Unit = {
-    val effects = basicUpdateStrategy.effects(BeforeExpression) ++ basicUpdateStrategy.effects(AfterExpression)
+  def syncWithStrategy(strategy: GenericUpdateStrategy): Unit = {
+    val effects = strategy.effects(BeforeExpression) ++ strategy.effects(AfterExpression)
     val lipSyncHolder = effects.find(_.isInstanceOf[LipSyncFromMotionSound]).map(_.asInstanceOf[LipSyncFromMotionSound])
     lipSync.setSelected(lipSyncHolder.isDefined)
     lipSyncHolder.foreach { effect =>